Changed /callback functionality. It no longer handles logic, Dashboard will do the API call from now on.
